Fix Windows MSVC CI: drop LLP64-redundant (size_t) casts in buffer IO helpers APX-184
clang-tidy readability-redundant-casting fails the Windows builds
(compiled with clang targeting the MSVC ABI, where size_t is
unsigned long long == u64): tests/integration/resource_object_buffer.c
cast its u64 byte-count parameters to size_t before calling
fs->write_file/fs->read_file, which is an explicit cast to the exact
same type on LLP64 targets. Remove the casts; the implicit u64 -> size_t
conversion is value-preserving on every supported platform (same-width
unsigned on LP64, identical type on LLP64) and matches the existing
pattern in core/resource_asset_builtins.c.

+++ tests/integration/resource_object_buffer.c
@@ -139,7 +139,11 @@ static i32 it_file_write_all(const_chr_t path, const void* data, u64 size) {
 		fs->remove(path);
 		return -1;
 	}
-	u64 written = fs->write_file(file, data, (size_t)size);
+	/* size_t is unsigned long long on Windows (LLP64), so an explicit
+	 * (size_t) cast would be redundant there; the implicit conversion is
+	 * value-preserving on every supported platform (matches
+	 * resource_asset_builtins.c). */
+	u64 written = fs->write_file(file, data, size);
 	fs->close_file(file);
 	if (written != size) {
 		fs->remove(path);
@@ -155,7 +159,7 @@ static i32 it_file_read_all(const_chr_t path, void* data, u64 size) {
 	if (file == NULL) {
 		return -1;
 	}
-	u64 got = fs->read_file(file, data, (size_t)size);
+	u64 got = fs->read_file(file, data, size);
 	fs->close_file(file);
 	return got == size ? 0 : -1;
 }
